% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q-and-a.R
\name{question}
\alias{question}
\title{Quiz questions}
\usage{
question(text, ...)
}
\arguments{
\item{text}{question or option text}

\item{...}{one or more questions or answers}
}
\value{

}
\description{
Add multiple choice quiz questions for quizlite. It is used within
  [quizlite()].
}
\examples{
question(text = "What is better?",
    answer("R", correct = TRUE),
    answer("Python"),
    answer("Javascript"),
    answer("S"))
}
