% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quizlite.R
\name{quizlite_knit}
\alias{quizlite_knit}
\title{Create knitted quiz}
\usage{
quizlite_knit(quiz_db, ...)
}
\arguments{
\item{quiz_db}{A nested list}

\item{...}{Extra functions. e.g. height or width for sizing}
}
\value{
A html quiz inside a knitted html_document
}
\description{
Create knitted quiz
}
